pkg/ead/component/: change CollectionChildren and CollectionLevel from `bool` and `int` to `string` [+]
This is prevent `panic` with error like "String method: Unrecognized `reflect.Type.Kind`: int)" in `getSolrFieldElementStringsInV1IndexerInsertionOrder()`.  Tried adding int and bool types to the if/then in that function, but the values didn't write out correctly.  It's not worth forcing those types because in the end everything ends up a string in XML anyway.