add util.GetRepoCode()
Add a utility function (and associated tests) that takes an EAD
filesystem path as a string and returns the EAD's archival repository
code "RepoCode" as a string.  Various tests are run on the EAD path
string to ensure that the input looks like a path to an EAD file in
one of the NYULibraries findingaids_eads Git repositories.