pkg/index/testsupport/gen-repo.bash: add step to rename branch [+]
Need to rename "main" -> "master".  Reasons:

- go-git uses "master" as a default when `CheckoutOptions.Branch` is not specified, and buggy behavior can result if that branch does not exist.
- The default branch of the EAD files repo(s) is "master", and we want to be as close to the real thing as possible.

For full details, see https://jira.nyu.edu/browse/DLFA-276.

The test git repo branch was already manually renamed in the previous commit.
The script change for this commit was tested by running it and confirming that the default branch was named correctly.
The test git repo was not replaced (no need).